# Organization Changes - WordleBot moved to directory to allow for substructure - Dictionaries moved out of WordleBot class file
# Guesser Class
Created a Guesser class to move guessing logic out of WordleBot class. Hopefully, will improve readability as we add strategies

#Entropy Guesser
New guessing strategy - calculate the expectation value of the number of words a guess will remove from the possible_words list for all possible guesses and then use that to rank guesses

#Light mode
Added light mode :D
